分布式master slave工作解耦完毕

+++ src/main/java/com/tl/job002/monitor/MonitorManager.java
@@ -60,7 +60,7 @@ public class MonitorManager {
 		return TaskScheduleManager.getDoneTaskSize();
 	}
 
-	public static int getTotalTodoUrlNumber() {
+	public static long getTotalTodoUrlNumber() {
 		return TaskScheduleManager.getTodoTaskSize();
 	}
